fix bug in CL methods
