+++ copo_code/copo/ccenv.py
@@ -1,4 +1,3 @@
-import copy
 from collections import defaultdict
 
 from metadrive.utils import norm
@@ -67,22 +66,3 @@ def get_ccenv(env_class):
     CCEnv.__name__ = name
     CCEnv.__qualname__ = name
     return CCEnv
-
-
-def get_change_n_env(env_class):
-    class ChangeNEnv(env_class):
-        def __init__(self, config):
-            self._raw_input_config = copy.deepcopy(config)
-            super(ChangeNEnv, self).__init__(config)
-
-        def close_and_reset_num_agents(self, num_agents):
-            config = copy.deepcopy(self._raw_input_config)
-            self.close()
-            config["num_agents"] = num_agents
-            super(ChangeNEnv, self).__init__(config)
-
-    name = env_class.__name__
-    name = "CL{}".format(name)
-    ChangeNEnv.__name__ = name
-    ChangeNEnv.__qualname__ = name
-    return ChangeNEnv
